DataManager was getting fat and ugly, introducing TeleportManager
