+++ src/main/java/net/syrupstudios/syrupessentials/util/DataManager.java
@@ -33,8 +33,6 @@ public class DataManager {
     private static final Logger LOGGER = LogUtils.getLogger();
     private final MinecraftServer minecraftServer;
     private static final Map<UUID, PlayerData> PLAYERS = new HashMap<>();
-    private static final HashMap<UUID, TeleportRequest> APPROVED_TELEPORTS = new HashMap<>();
-    private static final HashMap<UUID, TeleportRequest> TELEPORT_APPROVAL_REQUESTS = new HashMap<>();
     private WorldData worldData;
     private static long currentTick;
 
@@ -183,82 +181,10 @@ public class DataManager {
         return Optional.of(PLAYERS.computeIfAbsent(player.getUUID(), k -> new PlayerData(player.getUUID(), player.getGameProfile().getName())));
     }
 
-    public static int teleportRequest(ServerPlayer serverPlayer, MinecraftServer server, String destinationPlayerName){
-        ServerPlayer target = server.getPlayerList().getPlayerByName(destinationPlayerName);
-        if (target!=null) {
-            TELEPORT_APPROVAL_REQUESTS.put(
-                    serverPlayer.getUUID(),
-                    new TeleportRequest(
-                            new TeleportPos(target.level(), target.blockPosition(), target.getXRot(), target.getYRot()),
-                            target.getUUID(),
-                            serverPlayer,
-                            currentTick + 600 //replace with config timeout eventually
-                    ));
-            target.sendSystemMessage(Component.literal("Receiving teleport request from: "+serverPlayer.getDisplayName().getString()));
-            target.sendSystemMessage(Component.literal("Type /tpaccept or /tpdeny to respond.."));
-            return 1;
-        }
-        return -1;
-    }
-
-    public static int approveTeleportRequest(ServerPlayer receiver){
-        Optional<TeleportRequest> possibleTeleportRequest =
-                TELEPORT_APPROVAL_REQUESTS.values()
-                        .stream()
-                        .filter(tr -> tr.getReceiverPlayerUUID().equals(receiver.getUUID()))
-                        .findFirst();
-        if(possibleTeleportRequest.isPresent()){
-            possibleTeleportRequest.get().getSenderPlayer().sendSystemMessage(Component.literal("Teleport Request Approved, Teleporting.."));
-            receiver.sendSystemMessage(
-                    Component.literal(
-                            String.format("Teleport Request Approved, Teleporting %s to you..", possibleTeleportRequest.get().getSenderPlayer().getDisplayName().getString())));
-            TeleportRequest request = possibleTeleportRequest.get();
-            APPROVED_TELEPORTS.put(
-                    request.getSenderPlayer().getUUID(),
-                    request
-            );
-            TELEPORT_APPROVAL_REQUESTS.remove(receiver.getUUID());
-            return 1;
-        }
-        receiver.sendSystemMessage(Component.literal("Teleport Request Has Expired"));
-        return 0;
-    }
-
-    public static int denyTeleportRequest(ServerPlayer receiver){
-        Optional<TeleportRequest> possibleTeleportRequest =
-                TELEPORT_APPROVAL_REQUESTS.values()
-                        .stream()
-                        .filter(tr -> tr.getReceiverPlayerUUID().equals(receiver.getUUID()))
-                        .findFirst();
-        if(possibleTeleportRequest.isPresent()){
-            possibleTeleportRequest.get().getSenderPlayer().sendSystemMessage(Component.literal("Teleport Request Denied."));
-            receiver.sendSystemMessage(Component.literal("Teleport Request Denied."));
-            TELEPORT_APPROVAL_REQUESTS.remove(receiver.getUUID());
-            return 1;
-        }
-        receiver.sendSystemMessage(Component.literal("No currently pending teleports found."));
-        return 0;
-    }
-
     public void onServerTick(){
         currentTick++;
         if(currentTick % 600 == 0){
             savePlayers(minecraftServer);
         }
-        TELEPORT_APPROVAL_REQUESTS.entrySet().removeIf(
-                entry -> currentTick >= entry.getValue().getExpiresAtTick());
-
-        TELEPORT_APPROVAL_REQUESTS.entrySet().removeIf(
-                entry -> {
-                    if(currentTick >= entry.getValue().getExpiresAtTick()){
-                        entry.getValue().getSenderPlayer().sendSystemMessage(Component.literal("Teleport Request Timed Out."));
-                        return true;
-                    }
-                    return false;
-                }
-        );
-
-        APPROVED_TELEPORTS.entrySet().removeIf(entry ->
-                SyrupEssentials.teleportPlayer(entry.getValue().getTeleportPos(), entry.getValue().getSenderPlayer(), true));
     }
 }
